requirements.txt
flask
python-dotenv
flask-cors
Flask-Session
Werkzeug
urllib3>2
gevent
gunicorn